+++ src/py/kaleido/errors.py
@@ -1,6 +1,6 @@
 """A list of errors available from the kaleido package."""
 
-from .kaleido import JavascriptError, KaleidoError
+from ._kaleido_tab import JavascriptError, KaleidoError
 
 __all__ = [
     "JavascriptError",
